% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_result.R
\name{get_api_result}
\alias{get_api_result}
\title{Get API Result}
\usage{
get_api_result(search, start_record = 0)
}
\arguments{
\item{search}{A search instance.}

\item{start_record}{Sequence number of the first record to fetch, by default 0.}
}
\value{
A result from PubMed database.
}
\description{
This function returns results from PubMed database using the provided search parameters.
}
