% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_numeric_month_by_string}
\alias{get_numeric_month_by_string}
\title{Get Numeric Month by String}
\usage{
get_numeric_month_by_string(month_str)
}
\arguments{
\item{month_str}{A month name.}
}
\value{
The numeric equivalent of the month.
}
\description{
This function converts a month name to its numeric equivalent.
}
