% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper.R
\name{get_paper}
\alias{get_paper}
\title{Get Paper}
\usage{
get_paper(paper_entry, publication)
}
\arguments{
\item{paper_entry}{A paper entry retrieved from PubMed API.}

\item{publication}{A publication instance that will be associated with the paper.}
}
\value{
A paper instance or NULL.
}
\description{
Using a paper entry provided, this function builds a paper instance.
}
