% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paper_entry.R
\name{get_paper_entry}
\alias{get_paper_entry}
\title{Get Paper Entry}
\usage{
get_paper_entry(pubmed_id)
}
\arguments{
\item{pubmed_id}{A PubMed ID.}
}
\value{
A paper entry from PubMed database.
}
\description{
This function returns paper data from PubMed database using the provided PubMed ID.
}
