% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publication.R
\name{get_publication}
\alias{get_publication}
\title{Get Publication}
\usage{
get_publication(paper_entry)
}
\arguments{
\item{paper_entry}{A paper entry retrieved from PubMed API.}
}
\value{
A publication instance.
}
\description{
Using a paper entry provided, this function builds a publication instance.
}
