% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_url.R
\name{get_search_url}
\alias{get_search_url}
\title{Get Search URL}
\usage{
get_search_url(search, start_record = 0)
}
\arguments{
\item{search}{A search instance.}

\item{start_record}{Sequence number of the first record to fetch, by default 0.}
}
\value{
A URL to be used to retrieve data from PubMed database.
}
\description{
This function returns the URL to be used to retrieve data from PubMed database.
See https://www.ncbi.nlm.nih.gov/books/NBK25500/ for query tips.
}
