% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_recursively.R
\name{get_text_recursively}
\alias{get_text_recursively}
\title{Get Text Recursively}
\usage{
get_text_recursively(text_entry)
}
\arguments{
\item{text_entry}{An arbitrary object that contains some text.}
}
\value{
The extracted text.
}
\description{
This function extracts text from an arbitrary object recursively.
}
