% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmApiRequest.R
\name{pmApiRequest}
\alias{pmApiRequest}
\title{Gather bibliographic content from PubMed database using NCBI entrez API}
\usage{
pmApiRequest(query, limit, api_key = NULL)
}
\arguments{
\item{query}{is a character. It contains a search query formulated using the Entrez query language.}

\item{limit}{is numeric. It indicates the max number of records to download.}

\item{api_key}{is a character. It contains a valid api key API keys for the NCBI E-utilities.}
}
\value{
a list cointaining bibliographic metadata downloaded from NCBI PubMed.

To obtain a free access to NCBI API, please visit: \href{https://www.ncbi.nlm.nih.gov/pmc/tools/developers/}{https://www.ncbi.nlm.nih.gov/pmc/tools/developers/}

To obtain more information about how to wirte a NCBI search query, please visit: \href{}{}
}
\description{
It gathers metadata about publications from the NCBI PubMed database.
The function \code{pmApiRequest} queries NCBI PubMed using an entrez query formulated through the function \code{pmQueryBuild}.
}
\examples{

# query <- "bibliometric"
# D <- pmApiRequest(query = query, limit = 1000, api_key = NULL)

}
\seealso{
\code{\link{pmQueryBuild}}

\code{\link{pmQueryTotalCount}}

\code{\link{pmApi2df}}
}
