% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCitedBy.R
\name{pmCitedBy}
\alias{pmCitedBy}
\title{Find articles that cite a given PubMed article}
\usage{
pmCitedBy(pmid, api_key = NULL)
}
\arguments{
\item{pmid}{is a character or numeric. A single PubMed identifier (PMID).}

\item{api_key}{is a character. It contains a valid API key for the NCBI E-utilities.
Default is \code{api_key=NULL}. The API key can also be set via the environment variable
\code{PUBMED_API_KEY} or \code{ENTREZ_KEY}.}
}
\value{
a list containing:
\tabular{lll}{
pmid \tab \tab The queried PMID.\cr
cited_by \tab \tab A character vector of PMIDs that cite the queried article.\cr
count \tab \tab The number of citing articles found.}
}
\description{
It retrieves the PMIDs of articles that cite a given PubMed article,
using the NCBI E-Link service (PubMed Cited by).
}
\details{
This function uses the NCBI E-Link endpoint with linkname "pubmed_pubmed_citedin"
to find articles in PubMed that cite the given article.

Note: Citation data in PubMed is based on PubMed Central (PMC) and may not be as
comprehensive as commercial citation databases (e.g. Web of Science, Scopus).
}
\examples{

\donttest{
# Find articles that cite PMID 25824007
cites <- pmCitedBy(pmid = "25824007")
cites$count
cites$cited_by
}

}
\seealso{
\code{\link{pmReferences}}

\code{\link{pmFetchById}}
}
