% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmCitedBy.R
\name{pmReferences}
\alias{pmReferences}
\title{Find references cited by a given PubMed article}
\usage{
pmReferences(pmid, api_key = NULL)
}
\arguments{
\item{pmid}{is a character or numeric. A single PubMed identifier (PMID).}

\item{api_key}{is a character. It contains a valid API key for the NCBI E-utilities.
Default is \code{api_key=NULL}. The API key can also be set via the environment variable
\code{PUBMED_API_KEY} or \code{ENTREZ_KEY}.}
}
\value{
a list containing:
\tabular{lll}{
pmid \tab \tab The queried PMID.\cr
references \tab \tab A character vector of PMIDs referenced by the queried article.\cr
count \tab \tab The number of references found.}
}
\description{
It retrieves the PMIDs of articles that are cited by (referenced in) a given
PubMed article, using the NCBI E-Link service.
}
\details{
This function uses the NCBI E-Link endpoint with linkname "pubmed_pubmed_refs"
to find articles in PubMed that are referenced by the given article.

Note: Reference data is extracted from PubMed Central (PMC) full-text articles and
is only available when the full text is deposited in PMC. Not all PubMed articles
have reference data available.
}
\examples{

\donttest{
# Find references of PMID 25824007
refs <- pmReferences(pmid = "25824007")
refs$count
refs$references
}

}
\seealso{
\code{\link{pmCitedBy}}

\code{\link{pmFetchById}}
}
