% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedR-package.R
\docType{package}
\name{pubmedR-package}
\alias{pubmedR}
\alias{pubmedR-package}
\title{pubmedR: Fetching and Processing PubMed Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
