% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run PubMed Search}
\usage{
run(search)
}
\arguments{
\item{search}{A search instance.}
}
\description{
This function fetches papers from PubMed database using the provided search parameters.
After fetching the data from PubMed, the collected papers are added to the provided search instance.
}
