% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_success}
\alias{try_success}
\title{Try Success}
\usage{
try_success(expr, retries, pre_delay = 0)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{retries}{Number of retries.}

\item{pre_delay}{Delay before each retry.}
}
\value{
The result of the expression.
}
\description{
This function tries to evaluate an expression multiple times until it succeeds.
}
