% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmQueryTotalCount.R
\name{pmQueryTotalCount}
\alias{pmQueryTotalCount}
\title{Count the number of documents returned by a query
#'
It counts the number of documents that a query returns from the NCBI PubMed database.}
\usage{
pmQueryTotalCount(query, api_key = NULL)
}
\arguments{
\item{query}{is a character. It contains a search query formulated using the Entrez query language.
A query can be automatically generated using the function \code{pmQueryBuild}.}

\item{api_key}{is a character. It contains a valid api key API keys for the NCBI E-utilities.}
}
\value{
a list. It contains three objects:
\tabular{lll}{
n \tab  \tab the total number of records returned by the query\cr
query_translation \tab       \tab The query transaltion \cr
web_history \tab      \tab the web history object}

To obtain a free access to NCBI API, please visit: \href{https://www.ncbi.nlm.nih.gov/pmc/tools/developers/}{https://www.ncbi.nlm.nih.gov/pmc/tools/developers/}
}
\description{
Count the number of documents returned by a query
#'
It counts the number of documents that a query returns from the NCBI PubMed database.
}
\examples{

# query <- "bibliometric"
# D <- pmQueryTotalCount(query = query, api_key = NULL)

}
\seealso{
\code{\link{pmQueryBuild}}

\code{\link{pmApi2df}}
}
